requirements.txt
Django==3.2.25
matplotlib==3.5.3
Pillow==9.3.0
python-dotenv==0.21.0